requirements.txt
scrapy
beautifulsoup4
psycopg2
flask
pandas
python-dotenv